+++ db/migrations/20230627112201_create_table_orders.sql
@@ -1,13 +1,13 @@
 -- migrate:up
 CREATE TABLE orders (
-	id INT NOT NULL AUTO_INCREMENT,
-	user_id INT NOT NULL,
-	order_number INT NOT NULL,
-	total_price DECIMAL NOT NULL,
-	status_id INT NOT NULL,
-	created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	updated_at TIMESTAMP NULL ON UPDATE CURRENT_TIMESTAMP,
-	PRIMARY KEY (id)
+  id INT NOT NULL AUTO_INCREMENT,
+  user_id INT NOT NULL,
+  order_number INT NOT NULL,
+  total_price DECIMAL(10,2) NOT NULL,
+  status_id INT NOT NULL,
+  created_at TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  updated_at TIMESTAMP NULL ON UPDATE CURRENT_TIMESTAMP,
+  PRIMARY KEY (id)
 )
 
 -- migrate:down
